# Use raw encoding for numerical tensors