--- a/CHANGELOG.rst
+++ b/CHANGELOG.rst
@@ -12,7 +12,7 @@ Change log
 
 **New feature**
 
-- The instantiation time of constant numerical tensors with more than 100 elements has been significantly reduced by encoding their payload as raw bytes.
+- The instantiation time of constant numerical tensors has been significantly reduced by encoding their payload as raw bytes.
 
 
 0.15.0 (2025-06-06)