--- a/src/spox/_utils.py
+++ b/src/spox/_utils.py
@@ -43,42 +43,6 @@ def dtype_to_tensor_type(dtype_like: npt.DTypeLike) -> int:
         raise TypeError(err_msg)
 
 
-def from_array(arr: np.ndarray, name: str | None = None) -> onnx.TensorProto:
-    """Convert the given ``numpy.array`` into an ``onnx.TensorProto``.
-
-    As it may be useful to name the TensorProto (e.g. in
-    initializers), there is a ``name`` parameter.
-
-    This function differs from ``onnx.numpy_helper.from_array`` by not
-    using the ``raw_data`` field for small arrays with less then 100
-    elements.
-    """
-    # Using raw=True does not allow for meaningful inspection of the
-    # values in netron and elsewhere. However, creating large protobuf
-    # tensors with `raw=False` is an order of magnitude slower for
-    # large tensors. The following makes the compromise that large
-    # numerical arrays can be encoded in the raw format since they
-    # defy manual inspection anyhow. Creating the tensor is still
-    # rather involved and also requires handling of endianess. We thus
-    # still fall back to using `onnx.numpy_helper.from_array`, which
-    # uses raw=True under the hood.
-    if arr.size > 100 and arr.dtype not in [np.str_, np.object_]:
-        return onnx.numpy_helper.from_array(arr, name=name)
-    cast_to_bytes = False
-    if arr.dtype.type in [np.str_, np.object_]:
-        cast_to_bytes = True
-    return onnx.helper.make_tensor(
-        name=name or "",
-        data_type=dtype_to_tensor_type(arr.dtype),
-        dims=arr.shape,
-        # make_tensor fails on scalars. We fix it by calling flatten
-        vals=(
-            np.char.encode(arr, encoding="utf-8") if cast_to_bytes else arr
-        ).flatten(),
-        raw=False,
-    )
-
-
 def make_model(
     graph: onnx.GraphProto,
     *,