Protobuf tensor creation can be annoyingly slow for large tensors. This is particularly an issue for deep learning use cases and artifacts with large lookup tables. This change results in speedups of 10x or more (for the conversion time; runtime is unaffected). 

Running the following benchmark script:
```python
from time import perf_counter
import numpy as np

import spox.opset.ai.onnx.v21 as op

for size_exp in range(1, 9):
    size = 10**size_exp

    arr = np.ones((size,), dtype=np.float32)
    t0 = perf_counter()
    for _ in range(5):
        op.const(arr)
    t1 = perf_counter()

    print(f"{size=}: {(t1 - t0)}s")
```

yields the following results prior to this PR ("without raw") and after this PR:
```
# Without raw:

size=10: 0.003982624999480322s
size=100: 0.0008588329947087914s
size=1000: 0.0018544170015957206s
size=10000: 0.011751874990295619s
size=100000: 0.10788500000489876s
size=1000000: 1.087528750009369s
size=10000000: 11.295649749983568s
size=100000000: 122.66386904200772s

# With raw:

python bench.py
size=10: 0.001316792011493817s
size=100: 0.0008856659987941384s
size=1000: 0.0007787499926052988s
size=10000: 0.0015762920083943754s
size=100000: 0.0034534579899627715s
size=1000000: 0.0508717920165509s
size=10000000: 0.6581135420128703s
size=100000000: 12.970771457999945s
```

# Checklist

- [x] Added a `CHANGELOG.rst` entry