@cbourjau 2025-06-19
Prompted by your [comment](https://github.com/Quantco/spox/pull/235#discussion_r2156805200), I went ahead and tried things out (thanks!). Everything works as expected in Netron already! I dug into the archives to figure out why we had the current workaround in place. It's from the old, closed-source days and boils down to the readability/editability of the protobuf itself rather than anything Netron-related. That is not a good enough reason to keep the observed performance penalty, IMHO. I removed the override completely in favor of just using the upstream function.